Fix manpage-check on RHEL 5.
The version of groff on RHEL 5 doesn't include the .SY, .OP, or .YS macros
that ovs-benchmark.1 uses, so the manpage-check target fails on that
platform.  This commit adds the groff definitions of those macros to a
file and includes it into ovs-benchmark.1.

I tested that this allows RHEL 5 to pass manpage-check.