Do not output priority tagged frames.
Linux hosts (and probably others) tend to ignore priority-tagged frames, so
it's safest not to output them at all.

On "master" we've added a per-port setting for whether priority-tagged
frames may be output, but "master" and "branch-1.3" have diverged
significantly in how they handle port output, so it's much simpler to
port this simpler policy to "branch-1.3".

Reported-by: Michael Mao <mmao@nicira.com>
Bug #8320.