ofproto-dpif: Simplify output action composition.
Before this patch, the logic for outputting to a port was scattered
all around ofproto-dpif.  This patch simplifies the code by forcing
it to use one code path to check if a port is forwarding, and
output if appropriate.  Future patches will rely on this
simplification to implement new features.