ofproto-dpif: Restore optimization for no-actions case (without leak).
Commit 968131c1809 (ofproto-dpif: Omit "execute" operation entirely when
there are no actions.) introduced an optimization for the case where a
flow translated to ODP actions had no actions at all (i.e. the packet is
to be dropped).  It also introduced a memory leak (the packet was not
freed).

Commit 999fba59afd (ofproto-dpif: Implement PACKET_IN in userspace.)
inadvertently removed the optimization and as a side effect fixed the
memory leak.

This commit restores the optimization but not the memory leak.

Signed-off-by: Ben Pfaff <blp@nicira.com>